+++ Plugin/ConfigurableQuota/src/Metadata.cs
@@ -4,11 +4,13 @@ namespace ConfigurableQuota
     {
         public const string GUID = "com.seeya.configurablequota";
         public const string PLUGIN_NAME = "Configurable Quota";
-        public const string VERSION = "1.1.0";
+        public const string VERSION = "1.3.0";
 
         public const string LETHAL_NETWORK_API_GUID = "LethalNetworkAPI";
         public const string LETHAL_CONSTELLATIONS_GUID = "com.github.darmuh.LethalConstellations";
         public const string LETHAL_MOON_UNLOCKS_GUID = "com.xmods.lethalmoonunlocks";
         public const string OPEN_LIB_GUID = "darmuh.OpenLib";
+        public const string LLL_GUID = "imabatby.lethallevelloader";
+        public const string LUNAR_CONFIG_GUID = "Crafty.LunarConfig";
     }
 }